Fix investment support in management section
When supporting an investment in the management section through the
investment view, we were accessing an action in the public investments
controller. This meant the manager was the one supporting the investment
(as they'd be the `current_user` in this controller) and not the managed
user.

In the case of groups with many headings, voting the first time requires
a confirmation and then a regular (non-AJAX) request takes place. In
this case, users were redirected to the public area instead of remaining
in the management area.

Using the proper URL to vote solves the problem.

Note there was comment about one of these tests failing in Travis. Most
probably the test failed because there was no expectation after clicking
the link with the investment title, so the "Support" button (which is
also present in the index page) was clicked before the investment view
was loaded.